backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.3
numpy==1.26.4
pydantic==2.9.2
httpx==0.27.2
groq==0.11.0
python-dotenv==1.0.1
pytest==8.3.3
wbgapi==1.0.12
